backend/requirements.txt
fastapi[all]
SQLAlchemy
PyMySQL
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
